Increase RD cost comparison precision
master (-s=3) -> current (-s=0)

   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
-7.5800 |     N/A |     N/A |  -7.0594 | -7.8593 | -7.3970 |     4.0382

+++ src/rdo.rs
@@ -19,14 +19,14 @@ use std::vec::Vec;
 
 #[derive(Clone)]
 pub struct RDOOutput {
-    pub rd_cost: u64,
+    pub rd_cost: f64,
     pub part_type: PartitionType,
     pub part_modes: Vec<RDOPartitionOutput>
 }
 
 #[derive(Clone)]
 pub struct RDOPartitionOutput {
-    pub rd_cost: u64,
+    pub rd_cost: f64,
     pub bo: BlockOffset,
     pub pred_mode: PredictionMode
 }
